Fix bug with new metadata caching methodology and J2ME SD cards
Cache files generated by examining EPUBs used to be placed alongside
the content file itself. However Android does not allow writing to
the SD card except in a predetermined directory matching the app id.

A matching relative path is created underneath the main cache directory.
In J2ME this would include the drive letter and a colon - which must
be stripped out and was thus causing listing of SD card contents on
J2ME to fail.