Remove WiFi authentication failure handling
When a WiFi authentication error happened Android was already attempting to
connect to another network where available. As WiFi direct group connections
are done using the provided authentication credentials it is unlikely that this
would be a way to detect errors - we haven't seen it yet. We have kept logging
this event so it can be picked up on testing.

Refactor done on UMLogAndroid method.