Handle wifi disabled on supernode
1. When wifi is disabled and then re-enabled on the server group was not being formed correctly due to issues
both with WiFiBuddy and with how NetworkManagerAndroid was handling the wifi disabled event. This is now fixed

2. Tests are added to test when wifi is disabled both on the client and on the server. Unfortunately due to the buggy nature
of NSD on Android if the supernode has wifi disabled, then re-enabled whilst the NSD service is running on the client it
will not get picked up again. If the wifi is disabled on the supernode before starting the NSD service on the client it
will be picked up fine indicating there is likely no issue with restarting the NSD service on the supernode. There is a bug
with it being picked up again.