Change entry status task to use http only when node was discovered over network service discovery.
Devices could be connected to different networks. The network service discovery task was before using http if the other node had an advertised ip address and it did as well. If the IP address was unreachable, it would attempt to connect anyway, fail, and would then be unable to load the status of the desired entries.