Split DownloadJob into DownloadSet and DownloadJob to handle multiple download runs of any given set
DownloadSet entity is introduced to represent a given DownloadSet, each with one root entry (that is then to be shown on the Downloaded tab in the app)
A DownloadJob represents one run of downloading a particular DownloadSet. This run might be repeated for updating, or if the user deletes entries, and then changes their mind.