Refactor generation of Room DAOs and Database classes into an annotation processor that runs on core (part 1):
Previous system to generate Room DAOs and Database classes depended on an
executabletask that used reflection, rather than an actual annotation processor.

This version uses an annotation processor, with an argument for the output
directory (the annotation processor runs on core, but then outputs generated
source into the app-android module).