Update to ensure each db change mkaes a sequence number for each entity.
To ensure that the database sync can resume smoothly and be divided into
smaller parts, and resume from where it left off. If the same change seq
number is used for multiple entities, it becomes impossible to know for
the client to know if it has received the whole update.