Workaround memory leak bug in simple JNDI.
Simple JNDI database properties files must have a pool property to ensure a connection pool is used.
Apache commons DBCP2 must be added as a gradle dependency.

If a connection pool is not used, a new connection is created (and not disposed of) for every time
getConnection is called on the data source.