Fix Javascript database issue.
Sync logic was changed to ensure that an entity changed by a client is not sent back to the same client that changed it. This stopped Javascript clients from receiving an updated entity because the Javascript client id is always 0.

The query refactor for a syncable entity filter is now changed to deliver all entities to any client with an id of 0.

Signed-off-by: Mike Dawson <mike@ustadmobile.com>