Fix memory leaks on ContentEntryList and DownloadDialogPresenter.
ContentEntryList issue was caused by the wrapped live data holding a
reference to the parent object, which in turn held a reference to a view
(the repoloadingstatusview).

The wrapped LiveData is now based on a cleaner approach following the
same setup as MediatorLiveData (but using only one input source). Instead
of using a callback (which caused a memory leak) we provide a LiveData
object with the loading status which can be observed.

Signed-off-by: Mike Dawson <mike@ustadmobile.com>