Fix Epub XHTML filter. Fix table of contents navigation on epub
The Epub XHTML filter is applied to make content that responsive that is otherwise not responsive (e.g. epub
pages that do not have a meta viewport). This was not handling namespaces correctly.

Using the default recyclerview scrollToPosition does not reliably go to a page. This has been changed to use
scrollToPositionWithOffset and use .post call on the RecyclerView to ensure it is ready to accept this.