Create a normal DAO that contains various queries that are needed for sync helper entities. Do not implement DoorDatabaseSyncRepository in the generated SyncDao
Previously various queries that were needed for the helper entities that track sync (e.g.
UpdateNotification etc) were manually added by the annotation processor. There is now a normal
DAO which is easier to extend where these queries can be added as required.

The generated SyncDao Repo previously implemented DoorDatabaseSyncRepository - even though it
is not really the database. This had side effects for functions that were required for it to compile.
This has been removed.