Set edit screens fieldsEnabled uistate initial value to false.
Setting fieldsEnabled as true by default caused flakey behavior on UI tests: the component could be mounted to the DOM as enabled, which the viewmodel would then change to false, start loading, and set to true once loading was finished.

If the test caught the brief moment when fields were enabled and proceeded, then when it started typing, the fields could be disabled again.

Setting the initial value as false ensures that the components are not enabled when first mounted to the DOM, so when they are first enabled, it will be safe for a test to proceed.