Desktop PDF loader improvements and fix issue #828
Remove use of positioning measurement that caused display PDF to shrink on Windows. The desktop PdfFileHelper will now load the size of each page, and if a page is not ready, can create a placeholder image. This ensures the size of each item in the LazyColumn does not change, so pages will not jump.