Implement permission control on view course members and selecting a person to add.
Fix issue with approving enrolment request: student was not appearing in the students list after enrolment was approved because of a race condition: person entity was not in the local database, and selecting students is done first on the basis of the person entity. The server did not have the ClazzEnrolment when the invalidation happens, so it would not send the person entity.