Add template for bulk user addition and tidy up bulk user import UI
Fix edge case issue on GetStoragePathForUrlUseCaseCommonJvm: where the file was not compressed on the server (e.g. in staticfiles), but is compressed by the cache, the content encoding and size will not match. When checking for the size to ensure the file is ready, GetStoragePathForUrlUseCaseCommonJvm should compare to the cache storage size, not the content-length header from the original request.