Update report run/query to insert results into an intermediary table (ReportQueryResult).
Report runs can be large, complex, and run over millions of rows. They may also need to run on the server when not all underlying data is available on the client.

This commit updates report related use cases to store result rows in an intermediary table such that results can be cached server side and client side as needed.